HP/UX build problem An inconsistency between the class declaration and the implementation in the AOMF output filter has been fixed.
My thanks to R. op het Roodt <rroo@mail.com> for reporting this problem.

+++ include/srec/output/file/aomf.h
@@ -73,7 +73,7 @@ private:
       * Each has an 8-but type, a 16-bit little-endian length, a payload,
       * and an 8-bit 2s complement checksum.
       */
-    void emit_record(int, const unsigned char *, unsigned int);
+    void emit_record(int, const unsigned char *, size_t);
 
     /**
       * The module_header_record method is used to write an AOMF Module
